@@ -6,7 +6,7 @@ publish_to: 'none'
 
 environment:
   sdk: ">=3.0.0 <4.0.0"
-  flutter: ">=3.16.0"
+  flutter: ">=3.32.0"
 
 dependencies:
   flutter:
@@ -14,17 +14,23 @@ dependencies:
 
   cr_logger:
     path: ../
+  flutter_dropzone: 4.2.1
+  share_plus: ^12.0.1
 
-  dio: 5.4.0
-  chopper: 7.0.10
-  logger: 2.0.2+1
-  flutter_dropzone: 3.0.6
-  share_plus: 7.2.1
-
+  ferry: 0.16.1+2
+  gql_http_link: ^1.2.0
+  gql_code_builder_serializers: ^0.1.0+1
+  chopper: ^8.4.0
+  dio: ^5.9.0
+  logger: ^2.6.2
+  http: ^1.6.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  ferry_generator: ^0.12.0+2
+  build_runner: ^2.5.4
+  flutter_lints: ^6.0.0
 
 flutter:
   uses-material-design: true
