@@ -1,7 +1,7 @@
 name: cr_logger
 description: Powerful logging plugin. Supports android, ios and web platforms.
 
-version: 2.4.2
+version: 2.6.1
 homepage: https://github.com/Cleveroad/cr_logger
 
 topics:
@@ -13,73 +13,68 @@ topics:
 
 environment:
   sdk: ">=3.0.0 <4.0.0"
-  flutter: ">=3.16.0"
+  flutter: ">=3.32.0"
 
 dependencies:
   flutter:
     sdk: flutter
   flutter_web_plugins:
     sdk: flutter
-  http: ^1.1.2
 
+  # BSD-3-Clause
+  http: ">=0.13.0 <2.0.0"
   # MIT
   cr_json_widget: ^1.1.1
-
   # Network
   # MIT
-  dio: ^5.4.0
+  dio: ">=5.5.0+1 <6.0.0"
   # MIT
-  chopper: ^7.0.10
+  chopper: ">=8.0.0 <9.0.0"
   # MIT
-  logger: ^2.0.2+1
-
+  logger: ">=2.0.0 <3.0.0"
   # BSD
-  package_info_plus: ^5.0.1
-
+  package_info_plus: ">=8.0.0 <9.0.0"
   # MIT
-  uuid: ^3.0.7
-
+  uuid: ">=3.0.7 <5.0.0"
   # MIT
-  app_settings: ^5.1.1
-
+  app_settings: ">=5.0.0 <7.0.0"
   # BSD
-  path_provider: ^2.1.2
+  path_provider: ">=2.0.0 <2.2.0"
   # BSD-3-Clause
-  path: ^1.8.3
-
+  path: ">=1.8.0 <2.0.0"
   # MIT
-  inspector: ^2.1.0
-
+  inspector: ">=1.0.0 <4.0.0"
   # BSD-3-Clause
-  shared_preferences: ^2.2.2
+  shared_preferences: ">=2.0.0 <3.0.0"
   # MIT
   split_view: ^3.2.1
-
   # unknown
   # https://pub.dev/packages/synchronized/license
-  synchronized: ^3.1.0
-
+  synchronized: ">=3.0.0+2 <4.0.0"
   # BSD-3-Clause
-  js: ^0.6.7
-
-  # BSD-3-Clause
-  # https://stackoverflow.com/a/70757512
-  platform: ^3.1.2
-
+  js: ^0.7.2
   # Apache-2.0
-  infinite_scroll_pagination: ^4.0.0
+  infinite_scroll_pagination: ">=5.0.0 <6.0.0"
   # MIT
-  worker_manager: ^6.3.1
-
+  worker_manager: ">=7.0.0 <7.3.0"
   # BSD-2-Clause
-  sqflite: ^2.3.0
+  sqflite: ">=2.3.0 <3.0.0"
+  #MIT
+  gql_http_link: ">=1.0.1 <1.3.0"
+  #MIT
+  gql_link: ^1.1.0
+  #MIT
+  gql_exec: ^1.0.0+1
+  #MIT
+  gql: ^1.0.1
+  # BSD-3-Clause
+  web: ">=0.5.1 <2.0.0"
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^3.0.1
-
-  build_runner: ^2.4.7
+  # MIT
+  flutter_lints: ^6.0.0
 
 flutter:
   uses-material-design: true
