@@ -2,14 +2,19 @@ library cr_logger;
 
 export 'src/cr_logger.dart';
 export 'src/cr_logger_wrapper.dart';
-export 'src/data/bean/error_bean.dart';
+export 'src/data/bean/gql_bean.dart';
+export 'src/data/bean/graphql/graphql_error_bean.dart';
+export 'src/data/bean/graphql/graphql_request_bean.dart';
+export 'src/data/bean/graphql/graphql_response_bean.dart';
+export 'src/data/bean/http/http_error_bean.dart';
+export 'src/data/bean/http/http_request_bean.dart';
+export 'src/data/bean/http/http_response_bean.dart';
 export 'src/data/bean/http_bean.dart';
 export 'src/data/bean/log_bean.dart';
-export 'src/data/bean/request_bean.dart';
-export 'src/data/bean/response_bean.dart';
 export 'src/data/models/log_type.dart';
 export 'src/dio_log.dart';
 export 'src/extensions/ansi_color_ext.dart';
+export 'src/graphql/cr_logger_gql_link.dart';
 export 'src/managers/http_log_manager.dart';
 export 'src/page/actions_and_values/models/notifier_data.dart';
 export 'src/widget/cr_inspector.dart';
