@@ -1,8 +1,7 @@
 import 'package:flutter/material.dart';
 
 extension ImageExt on Image {
-  static Image fromPackage(
-    String name, {
+  static Image fromPackage(String name, {
     double width = 24,
     double height = 24,
   }) =>
