@@ -5,28 +5,46 @@ import 'package:chopper/chopper.dart';
 import 'package:cr_logger/cr_logger.dart';
 import 'package:http/http.dart' as http;
 
-final class ChopperLogInterceptor
-    implements ResponseInterceptor, RequestInterceptor {
+final class ChopperLogInterceptor implements Interceptor {
   final logManager = HttpLogManager.instance;
 
-  @override
-  FutureOr<Request> onRequest(Request request) async {
-    final reqOpt = RequestBean()
-      ..id = _getRequestHashCode(await request.toBaseRequest())
-      ..url = request.url.toString()
-      ..method = request.method
-      ..contentType = request.headers['Content-Type'].toString()
-      ..requestTime = DateTime.now()
-      ..params = request.parameters
-      ..body = request.body
-      ..headers = request.headers;
-    logManager.onRequest(reqOpt);
+  /// Creates hashcode based on request
+  int _getRequestHashCode(http.BaseRequest baseRequest) {
+    var hashCodeSum = 0;
 
-    return request;
+    hashCodeSum += baseRequest.url.hashCode;
+    hashCodeSum += baseRequest.method.hashCode;
+    if (baseRequest.headers.isNotEmpty) {
+      baseRequest.headers.forEach((key, value) {
+        hashCodeSum += key.hashCode;
+        hashCodeSum += value.hashCode;
+      });
+    }
+    if (baseRequest.contentLength != null) {
+      hashCodeSum += baseRequest.contentLength.hashCode;
+    }
+
+    return hashCodeSum.hashCode;
   }
 
   @override
-  FutureOr<Response> onResponse(Response<dynamic> response) {
+  FutureOr<Response<BodyType>> intercept<BodyType>(
+      Chain<BodyType> chain) async {
+    final request = chain.request;
+    final reqOpt = HttpRequestBean(
+      id: _getRequestHashCode(await request.toBaseRequest()),
+      url: request.url,
+      method: request.method,
+      contentType: request.headers['Content-Type'].toString(),
+      requestTime: DateTime.now(),
+      params: request.parameters,
+      body: request.body,
+      headers: request.headers,
+    );
+    logManager.onRequest(reqOpt);
+
+    final response = await chain.proceed(request);
+
     var data = <String, dynamic>{};
     try {
       data = json.decode(response.body as String);
@@ -51,48 +69,31 @@ final class ChopperLogInterceptor
       responseData = data.isNotEmpty ? data : response.body;
     }
 
-    final responseBean = ResponseBean()
-      ..id = requestId
-      ..responseTime = DateTime.now()
-      ..statusCode = response.statusCode
-      ..url = response.base.request?.url.toString()
-      ..method = response.base.request?.method
-      ..statusMessage = response.base.reasonPhrase
-      ..data = responseData
-      ..headers = response.headers;
-    logManager.onResponse(responseBean);
+    final responseBean = HttpResponseBean(
+      id: requestId,
+      responseTime: DateTime.now(),
+      statusCode: response.statusCode,
+      url: Uri(path: response.base.request?.url.toString()),
+      method: response.base.request?.method,
+      statusMessage: response.base.reasonPhrase,
+      data: responseData,
+      headers: response.headers,
+    );
+    await logManager.onResponse(responseBean);
 
     /// On error
     if (isError) {
-      final errorBean = ErrorBean()
-        ..id = requestId
-        ..errorData = response.error
-        ..statusCode = statusCode
-        ..statusMessage = response.base.reasonPhrase
-        ..url = response.base.request?.url.toString()
-        ..time = DateTime.now();
+      final errorBean = HttpErrorBean(
+        id: requestId,
+        errorData: response.error,
+        statusCode: statusCode,
+        statusMessage: response.base.reasonPhrase,
+        url: Uri(path: response.base.request?.url.toString()),
+        errorTime: DateTime.now(),
+      );
       logManager.onError(errorBean);
     }
 
     return response;
-  }
-
-  /// Creates hashcode based on request
-  int _getRequestHashCode(http.BaseRequest baseRequest) {
-    var hashCodeSum = 0;
-
-    hashCodeSum += baseRequest.url.hashCode;
-    hashCodeSum += baseRequest.method.hashCode;
-    if (baseRequest.headers.isNotEmpty) {
-      baseRequest.headers.forEach((key, value) {
-        hashCodeSum += key.hashCode;
-        hashCodeSum += value.hashCode;
-      });
-    }
-    if (baseRequest.contentLength != null) {
-      hashCodeSum += baseRequest.contentLength.hashCode;
-    }
-
-    return hashCodeSum.hashCode;
   }
 }
