@@ -3,7 +3,7 @@ import 'package:cr_logger/src/cr_logger_helper.dart';
 import 'package:cr_logger/src/js/console_output_worker.dart';
 import 'package:cr_logger/src/js/scripts.dart';
 import 'package:cr_logger/src/utils/html_stub.dart'
-    if (dart.library.js) 'dart:html' as html;
+if (dart.library.js) 'dart:html' as html;
 import 'package:flutter/foundation.dart';
 import 'package:logger/logger.dart';
 import 'package:worker_manager/worker_manager.dart';
@@ -20,7 +20,8 @@ final class ConsoleLogOutput extends LogOutput {
     await CRLoggerHelper.instance.lock.synchronized(() async {
       if (kIsWeb) {
         if (kReleaseMode || kProfileMode) {
-          final src = html.ScriptElement()..text = printLogsScript;
+          final src = html.ScriptElement()
+            ..text = printLogsScript;
           html.document.body?.append(src);
           printLogs(event.lines);
           src.remove();
@@ -39,7 +40,8 @@ final class ConsoleLogOutput extends LogOutput {
       ..id = kWorkerId
       ..text = workerScript;
     html.document.body?.append(srcWorker);
-    final srcCreateWorker = html.ScriptElement()..text = createWorkerScript;
+    final srcCreateWorker = html.ScriptElement()
+      ..text = createWorkerScript;
     html.document.body?.append(srcCreateWorker);
   }
 }
