@@ -8,17 +8,18 @@ Future<void> showInfoDialog({
 }) =>
     showDialog(
       context: context,
-      builder: (context) => Theme(
-        data: CRLoggerHelper.instance.theme,
-        child: AlertDialog(
-          title: title,
-          content: content,
-          actions: [
-            TextButton(
-              onPressed: () => Navigator.pop(context),
-              child: const Text('CLOSE'),
+      builder: (context) =>
+          Theme(
+            data: CRLoggerHelper.instance.theme,
+            child: AlertDialog(
+              title: title,
+              content: content,
+              actions: [
+                TextButton(
+                  onPressed: () => Navigator.pop(context),
+                  child: const Text('CLOSE'),
+                ),
+              ],
             ),
-          ],
-        ),
-      ),
+          ),
     );
