@@ -2,15 +2,17 @@ import 'package:cr_logger/src/utils/web_utils.dart';
 import 'package:cr_logger/src/widget/delete_log_confirm_widget.dart';
 import 'package:flutter/material.dart';
 
-Future<bool> showRemoveLogBottomSheet(
-  BuildContext context, {
+Future<bool> showRemoveLogBottomSheet(BuildContext context, {
   required String message,
   final Color textColor = Colors.black,
 }) async {
   final bottomSheetWidth =
-      MediaQuery.of(context).size.width > kWidthTrashHoldForMobileLayout
-          ? 400.0
-          : double.infinity;
+  MediaQuery
+      .of(context)
+      .size
+      .width > kWidthTrashHoldForMobileLayout
+      ? 400.0
+      : double.infinity;
 
   final result = await showModalBottomSheet<DeleteLogConfirmation?>(
     context: context,
@@ -21,10 +23,11 @@ Future<bool> showRemoveLogBottomSheet(
       ),
     ),
     constraints: BoxConstraints(maxWidth: bottomSheetWidth),
-    builder: (context) => DeleteLogConfirmWidget(
-      message: message,
-      textColor: textColor,
-    ),
+    builder: (context) =>
+        DeleteLogConfirmWidget(
+          message: message,
+          textColor: textColor,
+        ),
   );
 
   return result == DeleteLogConfirmation.ok;
